requirements.txt
# requirements.txt
Flask
requests
google-generativeai
gunicorn
python-dotenv